[chore] basic chat creation
