+++ app/src/main/res/values/strings.xml
@@ -64,4 +64,7 @@
     <string name="good">굿! 👍</string>
 
 
+    <string name="start_chat">티오와 대화 시작하기</string>
+
+
 </resources>
